# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market/quote/{symbol}:
    get:
      operationId: getQuote
      tags: [market]
      summary: Get a single live quote
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteDetail"
...
  /trading/orders:
    get:
      operationId: listOrders
      tags: [trading]
      summary: List orders for the current user
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [trading]
      summary: Place a buy or sell market order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /payments/deposit:
    post:
      operationId: createDepositCheckout
      tags: [payments]
      summary: Create a Stripe Checkout session to fund the account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepositBody"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateDepositResponse"
  /payments/confirm:
    post:
      operationId: confirmDeposit
      tags: [payments]
      summary: Confirm a completed checkout session and credit the account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmDepositBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSnapshot"
...
components:
  schemas:
...
    AccountSnapshot:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        cashBalance:
          type: number
        totalEquity:
          type: number
        portfolioValue:
          type: number
        dayChange:
          type: number
        dayChangePercent:
          type: number
        buyingPower:
          type: number
      required:
        - userId
        - displayName
        - cashBalance
        - totalEquity
        - portfolioValue
        - dayChange
        - dayChangePercent
        - buyingPower
...
    QuoteDetail:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        sector:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        previousClose:
          type: number
        open:
          type: number
        dayHigh:
          type: number
        dayLow:
          type: number
        yearHigh:
          type: number
        yearLow:
          type: number
        marketCap:
          type: number
        volume:
          type: number
        peRatio:
          type: number
        dividendYield:
          type: number
        description:
          type: string
      required:
        - symbol
        - name
        - sector
        - price
        - change
        - changePercent
        - previousClose
        - open
        - dayHigh
        - dayLow
        - yearHigh
        - yearLow
        - marketCap
        - volume
        - peRatio
        - dividendYield
        - description
...
    Order:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        price:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [filled, pending, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - symbol
        - name
        - side
        - quantity
        - price
        - total
        - status
        - createdAt
    PlaceOrderBody:
      type: object
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
      required:
        - symbol
        - side
        - quantity
...
    CreateDepositBody:
      type: object
      properties:
        amount:
          type: number
          description: USD deposit amount
      required:
        - amount
    CreateDepositResponse:
      type: object
      properties:
        url:
          type: string
        sessionId:
          type: string
      required:
        - url
        - sessionId
    ConfirmDepositBody:
      type: object
      properties:
        sessionId:
          type: string
      required:
        - sessionId
...