# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderSide } from "./orderSide";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  symbol: string;
  name: string;
  side: OrderSide;
  quantity: number;
  price: number;
  total: number;
  status: OrderStatus;
  createdAt: Date;
}